Create Team and Map structures
Teams can contain players and support a max number of players. If the
max players size is set to -1, it will allow an infinite number of
players.

Maps are objects that contain a World and a HashMap of locations for
the Teams. Maps can be unloaded without saving changes that were made
to them, meaning players can blow up the entire map while in a game and
it can be loaded again without having to copy the map.

Lombok was added as a dependency because it makes objects easier to
manage.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>