Adding Players, and starting and stopping games
Players can be added with Game#addPlayer, and they will be added to the
corrent Arena. Their inventory will be cleared and their health set to
full.

Starting an Arena will choose and random Map and teleport all the
players into it. Stopping an Arena will clear the players' inventories
and send them to the lobby.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>